northd: Use next-hop network for snat when lb_force_snat_ip=routerip.
When a gateway router has a load balancer configured, the option
lb_force_snat_ip=routerip can be set so that OVN snats load balanced
packets to the source logical router port's IP address, that is, the
port chosen as "outport" in the lr_in_ip_routing stage.

However, this was only designed to work when one network was configured
on the source logical router outport. When multiple networks are configured,
OVN's behavior was to simply choose the lexicographically first IP address for
snat. This lead to an incorrect address often being used for snat.

To fix this, two main components have been added:
 1. A new flag, flags.network_id. It is 4 bits and stores an index.
 2. A new stage in the router ingress pipeline, lr_in_network_id.

Now in the stage lr_in_network_id, OVN generates flows that assign
flags.network_id with an index. This index is then matched on later and the
network at that index will be chosen for snat.

Two tests have been added to verify that:
 1. The correct network is chosen for snat.
 2. The new and updated flows with flags.network_id are correct.

And tests that were broken by this new behavior have been updated.

Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>