controller: Prevent other chassis with same encap IP.
If you set a chassis' encap ip to be the same as that of another
chassis, OVN should detect it, log an error, and delete the chassis
record.

But this was not the case; OVN would allow the transaction. However, the
commit to the southbound database would fail, because there was already
an encap record with that same ip and type (from the other chassis).
The OVS db on the problematic chassis would then rapidly increase in
size. Essentially, the controller would get stuck in a loop of deleting
and recreating the tunnels.

This change makes it so when encaps are change, before building encaps,
the controller detects if there is already an existing encap tunnel of
the same ip and type. If so, it logs an error, refuses to build the
encaps, and deletes the associated chassis and chassis private record.

TO-DO:
 - [ ] Add case in test where chassis is deleted, and then add back in
       the check that it's disappeared.
 - [ ] Update commit message to explain why not running encaps if a
       transaction is in progress, maybe add a comment too.

Reported-at: https://issues.redhat.com/browse/FDP-1481
Co-authored-by: Nicholas Hubbard <nicholashubbard@posteo.net>
Signed-off-by: Nicholas Hubbard <nicholashubbard@posteo.net>
Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>