controller-vtep: Do not double add vtep ls.
One shash was being used to store the names of vtep logical switches AND to
see if there were any conflicts (i.e. X). Inside of check_pb_conflict(),
the vtep ls was being added again if

Removal from a shash is only dependent on the name of the node, not the
data it stores. It also removes the first match found, and that is
dependent on the order the hashmap stores nodes in, so it's not safe to
remove like this.

Fixes: INSERT COMMIT HERE, BLAME
I might have messed this up a little merging

Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>